Create header, footer, arrays and nav files
Created a header and footer file as entire website will use the same header and footer, this is to remove duplicate code. Each page will have its own content on it. Have created a new file for the nav bar so it will now run a foreach loop for each item in the navigation bar, to add or remove items this is done in the arrays file