added about page and changed header image
Added an about page and changed the header image because I got sick of looking at the burrito, also changed the height of the image as it looked a bit stretched. Made the years that it is open dynamic so it will +1 to the years for each year it has been open.